requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.5
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-guardian==1.4.9
django-redis==4.8.0
django-reversion==2.0.10
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.0.4
olefile==0.44
Pillow==4.2.1
pycryptodome==3.4.7
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.2
raven==6.2.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
uritemplate==3.0.0
urllib3==1.22
XlsxWriter==1.0.0
xlwt==1.3.0
